+++ R/em_profile_c.R
@@ -37,7 +37,8 @@ em_profile_c <- function(y, x_obj, par, c_candidates) {
   loglik <- as.numeric(log_lik_profile_fun(
     par$Beta.multinom.ZC, par$Beta.multinom.PL, par$Beta.NB, par$Alpha.NB,
     par$Beta.PL, c_candidates,
-    ext$zc, ext$pl_mult, ext$nb, ext$pl, y, ext$offset
+    ext$zc, ext$pl_mult, ext$nb, ext$pl, y, ext$offset,
+    ext$offset_zc, ext$offset_pl_mult
   ))
 
   # audit0.10 §1.4: which(loglik == max(loglik)) returns a vector on an exact