+++ R/parallel.R
@@ -191,6 +191,8 @@ evinf_boot_spec <- function(full_run) {
     x.multinom.pl = full_run$data$x.multinom.pl
   )
   spec$offset_nb <- full_run$offset_nb
+  spec$offset_zc <- full_run$offset_zc
+  spec$offset_pl_mult <- full_run$offset_pl_mult
   spec$control   <- full_run$control
   spec$coef      <- full_run$coef
   spec$formulas  <- full_run$formulas